Relocation of stuff into the API package - Also buildscript update
