+++ src/main/java/com/sk89q/worldedit/forge/CommonProxy.java
@@ -20,14 +20,14 @@
 package com.sk89q.worldedit.forge;
 
 import com.sk89q.worldedit.forge.gui.GuiHandler;
-import com.sk89q.worldedit.forge.network.WENetWrapper;
+import com.sk89q.worldedit.forge.network.WENetHandler;
 import cpw.mods.fml.common.network.NetworkRegistry;
 
 public class CommonProxy {
 
     public void registerHandlers() {
         NetworkRegistry.INSTANCE.registerGuiHandler(ForgeWorldEdit.inst, new GuiHandler());
-        WENetWrapper.init(ForgeWorldEdit.inst.getConfig());
+        WENetHandler.init(ForgeWorldEdit.inst.getConfig());
     }
 
 }
